year3/SWE/ngUML.component.backend/nguml/usecase/requirements.txt
aniso8601==9.0.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
blis==0.7.8
catalogue==2.0.8
certifi==2022.9.24
cffi==1.15.1
channels==3.0.5
charset-normalizer==2.1.1
click==8.1.3
confection==0.0.2
constantly==15.1.0
cryptography==38.0.1
cymem==2.0.6
daphne==3.0.2
Deprecated==1.2.13
Django==3.2.15
django-clearcache==1.2.1
django-cors-headers==3.13.0
django-debug-toolbar==3.7.0
django-filter==22.1
django-graphiql-debug-toolbar==0.2.0
django-js-asset==2.0.0
django-mptt==0.14.0
django-redis==5.2.0
django-rest-framework==0.1.0
django-rq==2.5.1
django-schema-graph==2.1.0
django-tables2==2.4.1
django-taggit==3.0.0
django-timezone-field==5.0
djangorestframework==3.14.0
drf-nested-routers==0.93.4
drf-spectacular==0.24.2
ghp-import==2.1.0
graphene==3.1.1
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
hyperlink==21.0.0
idna==3.4
incremental==21.3.0
inflection==0.5.1
Jinja2==3.1.2
jsonschema==4.16.0
langcodes==3.3.0
Markdown==3.3.7
MarkupSafe==2.1.1
mergedeep==1.3.4
mkdocs==1.4.0
mkdocs-autorefs==0.4.1
mkdocs-material==8.5.6
mkdocs-material-extensions==1.0.3
mkdocstrings==0.19.0
murmurhash==1.0.8
numpy==1.23.3
packaging==21.3
pathy==0.6.2
preshed==3.0.7
promise==2.3
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.2
Pygments==2.13.0
pymdown-extensions==9.6
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.4
PyYAML==6.0
pyyaml_env_tag==0.1
redis==4.3.4
requests==2.28.1
rq==1.11.1
service-identity==21.1.0
six==1.16.0
smart-open==5.2.1
spacy==3.4.1
spacy-legacy==3.0.10
spacy-loggers==1.0.3
sqlparse==0.4.3
srsly==2.4.4
text-unidecode==1.3
thinc==8.1.2
tqdm==4.64.1
Twisted==22.8.0
txaio==22.2.1
typer==0.4.2
typing_extensions==4.3.0
tzdata==2022.4
uritemplate==4.1.1
urllib3==1.26.12
wasabi==0.10.1
watchdog==2.1.9
wrapt==1.14.1
zope.interface==5.4.0
